+++ src/main/java/Memory/Types/Data.java
@@ -1,0 +1,7 @@
+package Memory.Types;
+
+import org.jdom2.Element;
+
+public abstract class Data {
+    public abstract Element getElement();
+}