+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 - [v1.1.0] Added Title Card (#3) (25 Nov 23)
   - [v1.1.1] Minor Problem Fixes (25 Nov 23)
   - [v1.1.2] Add Dev Route Picker (25 Nov 23)
-- [v1.2.0] Add Memory Pacakge (25 Nov 23)
+- [v1.2.0] Add Memory Package (25 Nov 23)
   - [v1.2.1] Organize Classes (25 Nov 23)
+  - [v1.2.2] Add Memory Creation and Recreation (25 Nov 23)
 
 ## v2.0.0 *The Usable Edition*